+++ src/main/groovy/org/folio/inventory/domain/Item.groovy
@@ -9,7 +9,8 @@ class Item {
   final String materialTypeId
   final String permanentLoanTypeId
   final String temporaryLoanTypeId
-  final String location
+  final String permanentLocationId
+  final String temporaryLocationId
 
   Item(String id,
        String title,
@@ -17,7 +18,8 @@ class Item {
        String instanceId,
        String status,
        String materialTypeId,
-       String location,
+       String permanentLocationId,
+       String temporaryLocationId,
        String permanentLoanTypeId,
        String temporaryLoanTypeId) {
 
@@ -27,21 +29,22 @@ class Item {
     this.instanceId = instanceId
     this.status = status
     this.materialTypeId = materialTypeId
-    this.location = location
+    this.permanentLocationId = permanentLocationId
+    this.temporaryLocationId = temporaryLocationId
     this.permanentLoanTypeId = permanentLoanTypeId
     this.temporaryLoanTypeId = temporaryLoanTypeId
   }
 
   Item copyWithNewId(String newId) {
     new Item(newId, this.title, this.barcode,
-      this.instanceId, this.status, this.materialTypeId, this.location,
-      this.permanentLoanTypeId, this.temporaryLoanTypeId)
+      this.instanceId, this.status, this.materialTypeId, this.permanentLocationId,
+      this.temporaryLocationId, this.permanentLoanTypeId, this.temporaryLoanTypeId)
   }
 
   def changeStatus(String newStatus) {
     new Item(this.id, this.title, this.barcode,
-      this.instanceId, newStatus, this.materialTypeId, this.location,
-      this.permanentLoanTypeId, this.temporaryLoanTypeId)
+      this.instanceId, newStatus, this.materialTypeId, this.permanentLocationId,
+      this.temporaryLocationId, this.permanentLoanTypeId, this.temporaryLoanTypeId)
   }
 
   @Override